LayoutTests/accessibility/placeholder.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
<script src="../resources/accessibility-helper.js"></script>
</head>
<body>

<input placeholder="search" type="text" name="q" size="15" maxlength="255" id="searchterm" />

<input id="password" class="field" type="password" name="sc1798" value="" placeholder="Password" spellcheck="true" maxlength="5096">

<span id="label">Birthday:</span>
<div id="search" role="searchbox" aria-labelledby="label" aria-placeholder="MM-DD-YYYY">03-14-1879</div>

<input type="text" id="input" placeholder="Fill in the blank" aria-placeholder="aria placeholder">
    
<script>
    description("This test makes sure that the placeholder is returned as the correct attribute");

    var fieldElement, passwordElement, searchElement, textInput;
    if (window.accessibilityController) {
        window.jsTestIsAsync = true;

        setTimeout(async function() {
            fieldElement = await waitForElementById("searchterm");
            shouldBe("fieldElement.stringAttributeValue('AXPlaceholderValue')", "'search'");

            passwordElement = await waitForElementById("password");
            shouldBe("passwordElement.stringAttributeValue('AXPlaceholderValue')", "'Password'");
            
            searchElement = await waitForElementById("search");
            shouldBe("searchElement.stringAttributeValue('AXPlaceholderValue')", "'MM-DD-YYYY'");
            
            // When the placeholder and aria-placeholder attributes are both present, use the placeholder
            // attribute's value.
            textInput = await waitForElementById("input");
            shouldBe("textInput.stringAttributeValue('AXPlaceholderValue')", "'Fill in the blank'");

            finishJSTest();
        }, 0);
    }
</script>
</body>
</html>
