LayoutTests/accessibility/placeholder-expected.txt

This test makes sure that the placeholder is returned as the correct attribute

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS fieldElement.stringAttributeValue('AXPlaceholderValue') is 'search'
PASS passwordElement.stringAttributeValue('AXPlaceholderValue') is 'Password'
PASS searchElement.stringAttributeValue('AXPlaceholderValue') is 'MM-DD-YYYY'
PASS textInput.stringAttributeValue('AXPlaceholderValue') is 'Fill in the blank'
PASS successfullyParsed is true

TEST COMPLETE
   Birthday:
03-14-1879
